+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWKYBAVSS6REXPM85E0XM3VT"
+        "01KWM4SKNNQ8A7A4J535T6QY1S"
     )
     r"""The unique identifier of the data source"""