+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-03T12:12:55.302Z"
+    created: Optional[str] = "2026-07-03T14:05:30.205Z"
 
-    updated: Optional[str] = "2026-07-03T12:12:55.303Z"
+    updated: Optional[str] = "2026-07-03T14:05:30.205Z"
 
     updated_by_id: OptionalNullable[str] = UNSET