+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-03T12:12:54.165Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-03T14:05:29.032Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")